moved core elements inside src/core
