+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import configureStore from './store'
+import configureStore from './core/store'
 import App from './App'
 
 const store = configureStore()
