+++ src/styles/main.scss
@@ -1,27 +1,32 @@
-@import 'ui/components/all';
-@import 'vars';
+@import "ui/components/all";
+@import "vars";
 
-*
-{
+* {
   font-size: 1em;
   font-family: Roboto !important;
 }
 
-html, body, .app, #root, [data-reactroot] {
-  height: 100%;
-  background-color: $bg-color;
-  margin:0 !important;
-  padding:0 !important;
-  position: relative;
-  text-align: center;
+html,
+body,
+.app,
+#root,
+[data-reactroot] {
+  top: 0 !important;
+  position: absolute !important;
+  width: 100% !important;
+  height: 100% !important;
+  margin: 0 !important;
+  padding: 0 !important;
+  text-align: center !important;
+  background-color: $bg-color !important;
 }
 
-.game{
+.game {
   display: flex;
   justify-content: space-around;
 }
 
-h1{
+h1 {
   color: ghostwhite;
   font-size: 4em;
 }
