+++ src/styles/ui/components/_all.scss
@@ -1,3 +1,4 @@
 @import "chatbox";
 @import "character";
 @import "login";
+@import "loading"
